Moved test helpers out of unit_test dir to top level test dir
