+++ tests/helpers/analytics.py
