+++ tests/integration_tests/collection/test_update.py
@@ -7,8 +7,8 @@ from fastapi import status
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 
+from tests.helpers.collection import create_collection_write_dto
 from tests.integration_tests.setup_db import EXPECTED_COLLECTIONS, setup_db
-from tests.unit_tests.helpers.collection import create_collection_write_dto
 
 
 def test_update_collection(client: TestClient, test_db: Session, user_header_token):
