+++ tests/integration_tests/document/test_create.py
@@ -6,8 +6,8 @@ from fastapi import status
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 
+from tests.helpers.document import create_document_create_dto
 from tests.integration_tests.setup_db import setup_db
-from tests.unit_tests.helpers.document import create_document_create_dto
 
 
 def test_create_document(client: TestClient, test_db: Session, user_header_token):
