+++ tests/integration_tests/event/test_create.py
@@ -4,8 +4,8 @@ from fastapi.encoders import jsonable_encoder
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import Session
 
+from tests.helpers.event import create_event_create_dto
 from tests.integration_tests.setup_db import setup_db
-from tests.unit_tests.helpers.event import create_event_create_dto
 
 
 def test_create_event(client: TestClient, test_db: Session, user_header_token):
