+++ tests/unit_tests/mocks/repos/collection_repo.py
@@ -5,7 +5,7 @@ from sqlalchemy import exc
 
 from app.errors import RepositoryError
 from app.model.collection import CollectionReadDTO
-from tests.unit_tests.helpers.collection import create_collection_read_dto
+from tests.helpers.collection import create_collection_read_dto
 
 ORG_ID = 1
 INCORRECT_ORG_ID = 1234
