+++ tests/unit_tests/routers/test_event.py
@@ -5,10 +5,7 @@ from fastapi import status
 from fastapi.encoders import jsonable_encoder
 from fastapi.testclient import TestClient
 
-from tests.unit_tests.helpers.event import (
-    create_event_create_dto,
-    create_event_write_dto,
-)
+from tests.helpers.event import create_event_create_dto, create_event_write_dto
 
 
 def test_get_all_when_ok(client: TestClient, user_header_token, event_service_mock):
